+++ helpme/src/main/java/com/developers/repository/RoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.developers.model.Case;
 
 @Repository
-public interface RoleRepository extends CrudRepository<Case, Integer> {
+public interface RoleRepository extends CrudRepository<Case, Long> {
 
 }
